Create Read-me
In this project, we are developing a simple Chat application. TCP Client and Server programs are executed for communication between resources using GENI.

In Question 1, we create a chat server which accepts a single client connection and displays everything the client types and the communication get terminated from both sides when the client types ‘exit’.

In Question 2, extending the Question 1, when one client terminates, the server remains open for the other client’s other clients that wish to communicate and can handle only one connection at a time.

In Question 3, extending the Question 2, the server will be able to handle multiple clients at a time and the server window displays the messages from the clients.

In Question 4, extending the Question 3, the server even echoes the messages from a client to all the remaining clients in the network.